fixed bugs in circuit4 example, which includes uncontracted indices for both tensor1 and tensor2.
Also adding DEBUG macro to switch off print statements.

Contraction output index was not being calculated correctly in this case due to bugs in getContractionVqIndices and the fact that the ordering of virtual qubits wasn't being considered.

VqVertex lists were not being updated correctly -- the link between tensor1 vertices and tensor2 vertices was not established correctly.